Метрики теперь являются вещественными (#206)
Предполагается использовать наносекундное разрешение, поэтому для
совместимости с 32-разрядными платформами используем double —
52 двоичных разряда (целые числа до 2↑52 даблами считаются точно).

+++ src/srlib/refalrts-allocator.cpp
@@ -131,7 +131,6 @@ void refalrts::Allocator::Pool::free() {
   }
 }
 
-void refalrts::Allocator::read_counters(unsigned long counters[]) {
-  counters[cPerformanceCounter_HeapSize] =
-    static_cast<unsigned long>(memory_use() * sizeof(Node));
+void refalrts::Allocator::read_counters(double counters[]) {
+  counters[cPerformanceCounter_HeapSize] = memory_use() * double(sizeof(Node));
 }
